fastapi_server/requirements.txt
# FastAPI & Web
fastapi[all]==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic-settings==2.6.1

# ML & Career Recommendation
scikit-learn==1.6.0
pandas==2.2.3
numpy==2.2.1
joblib==1.4.2
pydantic==2.10.3

# CLIP & Image Processing
torch>=2.0.0
torchvision>=0.15.0
clip-by-openai>=1.0
Pillow>=9.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Utilities
requests>=2.28.0
python-dotenv>=1.0.0

# Development (optional)
pytest>=7.0.0
httpx>=0.24.0
